refactoring
move adapter package to transfer
create extra package for ParameterCollection test classes
