+++ src/main/java/com/github/error418/opennms/client/transfer/adapter/DateAdapter.java
@@ -1,4 +1,4 @@
-package com.github.error418.opennms.client.adapter;
+package com.github.error418.opennms.client.transfer.adapter;
 
 import java.text.DateFormat;
 import java.util.Date;
