+++ src/test/java/com/github/error418/opennms/client/parameter/collection/CustomParameterCollection.java
@@ -1,6 +1,7 @@
-package com.github.error418.opennms.client.parameter;
+package com.github.error418.opennms.client.parameter.collection;
 
 import com.github.error418.opennms.client.exception.ParameterValidationException;
+import com.github.error418.opennms.client.parameter.ParameterCollection;
 
 /**
  * A custom Parameter Collection.
@@ -8,9 +9,9 @@ import com.github.error418.opennms.client.exception.ParameterValidationException
  */
 public class CustomParameterCollection extends ParameterCollection {
 
-	final static String FIRST_PARAM = "firstParameter";
-	final static String SECOND_PARAM = "secondParameter";
-	final static String THIRD_PARAM = "thirdParameter";
+	public final static String FIRST_PARAM = "firstParameter";
+	public final static String SECOND_PARAM = "secondParameter";
+	public final static String THIRD_PARAM = "thirdParameter";
 
 	private String[] parameterOrder = { FIRST_PARAM, SECOND_PARAM, THIRD_PARAM };
 
@@ -19,7 +20,7 @@ public class CustomParameterCollection extends ParameterCollection {
 			this.configureParameter(parameterName);
 		}
 	}
-	
+
 	public CustomParameterCollection setFirstParameter(String value) {
 		this.setValue(FIRST_PARAM, value);
 		return this;
@@ -37,7 +38,7 @@ public class CustomParameterCollection extends ParameterCollection {
 
 	@Override
 	protected void validate() throws ParameterValidationException {
-		if(this.isParameterNull(FIRST_PARAM)) {
+		if (this.isParameterNull(FIRST_PARAM)) {
 			throw new ParameterValidationException();
 		}
 	}
