+++ src/test/java/com/github/error418/opennms/client/parameter/collection/DefaultValueParameterCollection.java
@@ -1,6 +1,7 @@
-package com.github.error418.opennms.client.parameter;
+package com.github.error418.opennms.client.parameter.collection;
 
 import com.github.error418.opennms.client.exception.ParameterValidationException;
+import com.github.error418.opennms.client.parameter.ParameterCollection;
 
 public class DefaultValueParameterCollection extends ParameterCollection {
